pyproject.toml
[tool.poetry]
name = "nexus-conversations"
version = "0.1.0"
description = "Conversation Microservice for Nexus AI - Internal service for processing conversation messages from SQS FIFO queue"
authors = ["Ruan Rocha <ruan.rocha@vtex.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
django = "^4.2.6"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
celery = "5.3.6"
django-celery-beat = "2.5.0"
django-celery-results = "2.5.1"
redis = "^5.0.1"
django-redis = "^5.4.0"
hiredis = "^3.0.0"
boto3 = "^1.39.4"
sentry-sdk = "^1.39.2"
urllib3 = "^2.0.0"
pendulum = "^2.1.2"
weni-datalake-sdk = "0.2.2"
gunicorn = "^21.2.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
django-filter = "^23.5"

[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"
pytest-django = "^4.6.0"
pytest-cov = "^4.1.0"
ruff = "^0.6.0"
isort = "^5.12.0"
blue = "^0.9.1"
freezegun = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
exclude = ['.venv', 'migrations', 'manage.py', 'wsgi.py', 'asgi.py', '__init__.py']

[tool.ruff.lint]
select = [
    "E",        # pycodestyle (Error)
    "W",        # pycodestyle (Warning)
    "F",        # Pyflakes
    "B",        # flake8-bugbear
    "Q",        # flake8-quotes
    "UP",       # pyupgrade
    "I",        # isort
    "C901",     # mccabe (complex-structure)
    "DJ",       # flake8-django
]
ignore = [
    "DJ001"     # django-nullable-model-string-field
]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.isort]
profile = "black"
line_length = 120
extend_skip = ['migrations', 'manage.py', 'wsgi.py', 'asgi.py', '__init__.py']

[tool.pytest.ini_options]
pythonpath = "."
python_files = "tests.py test_*.py *_tests.py"
DJANGO_SETTINGS_MODULE = "nexus_conversations.settings"
addopts = "--nomigrations"
norecursedirs = [
  "staticfiles"
]

[tool.coverage.report]
omit = [
  "settings.py",
  "asgi.py",
  "wsgi.py",
  "__init__.py",
  "*/migrations/*",
  "manage.py",
  "urls.py",
  "*/tests/*",
  "*/test_*.py",
  "*/conftest.py",
  "*/main.py",
  "*/consumers/*",
  "*/exceptions.py",
]
